+++ element/30_tab_base.js
@@ -1,0 +1,29 @@
+/**
+ * The base class for tab elements
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+const TabBase = Function.inherits('Alchemy.Element.Form.Base', 'TabBase');
+
+/**
+ * Don't register this as a custom element,
+ * but don't let child classes inherit this
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabBase.makeAbstractClass();
+
+/**
+ * Get the context element
+ *
+ * @author   Jelle De Loecker <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabBase.setProperty(function tab_context() {
+	return this.queryUp('al-tab-context');
+});