+++ element/al_tab_panel.js
@@ -1,0 +1,44 @@
+/**
+ * The alchemy-tab-content element
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+const TabPanel = Function.inherits('Alchemy.Element.Form.TabBase', 'TabPanel');
+
+/**
+ * Set the ARIA role of the element
+ *
+ * @author   Jelle De Loecker <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabPanel.setRole('tabpanel');
+
+/**
+ * Make it tabable by default
+ *
+ * @author   Jelle De Loecker <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabPanel.setAttribute('tabindex', {default: 0});
+
+/**
+ * The name of the tab
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabPanel.setAttribute('tab-name');
+
+/**
+ * Is this tab active?
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.0
+ * @version  0.2.0
+ */
+TabPanel.setAttribute('active', {boolean: true});